% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{austr}
\alias{austr}
\title{Data set for testing marima package (australian killings)}
\format{A data frame (austr) with 7 columns and 100 rows.

\describe{
  \item{Year}{Year for data}
  \item{suic.fire}{Rate of suicides by firearms}
  \item{homi.fire}{Rate of homicides by firearms}
  \item{suic.other}{Rate of suicides by non firearms}
  \item{homi.other}{Rate of homicides by non firearms}
  \item{leg}{Legislation against firearms in effect}
  \item{acc.elg}{Accumulated effect of legislation in years} 
}}
\usage{
data(austr)
}
\description{
Data for marima examples.
}

