% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.polyn.R
\name{check.one}
\alias{check.one}
\title{check.one}
\usage{
check.one(polyn = NULL)
}
\arguments{
\item{polyn}{(k, k, ...) matrix polynomium with or without leading
unity matrix.}
}
\value{
polyn (array) with a leading unity matrix being
inserted if not present.
}
\description{
Function to check and insert leading unity matrix
if NOT present.
}
\examples{
set.seed(4711)
X <- array(rnorm(32),dim=c(4, 4, 2))
X <- check.one(X)
short.form(X)

}

