% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.marima.R
\name{print.marima}
\alias{print.marima}
\title{print.marima}
\usage{
\method{print}{marima}(x, estimates = TRUE, pvalues = FALSE,
  pattern = TRUE, fvalues = TRUE, ...)
}
\arguments{
\item{x}{= a marima object with results of marima analysis.}

\item{estimates}{= TRUE/FALSE: printout of parameter estimates.}

\item{pvalues}{= TRUE/FALSE: printout of (approximate) p-values for
parameter estimates.}

\item{pattern}{= TRUE/FALSE: printout of model definition pattern(s).}

\item{fvalues}{= TRUE/FALSE: printout of parameter (approximate)
F-values.}

\item{...}{Not used.}
}
\description{
Print some (most relevant) content of a marima object.
}

