% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lin.Results.R
\name{Results}
\alias{Results}
\title{Results}
\usage{
Results(Model = NULL)
}
\arguments{
\item{Model}{= an 'lm' object (with NA's for non-identifiable
ar.parameters in the 'lm' specification, if so).}
}
\value{
Results = matrix with 4 columns containing the same information
as summary(Model),
but with "NA" rows replaced by rows = c(0, 0, 0, NaN).
}
\description{
function which generates a matrix from summary(Model),
where 'Model' is an 'lm' object. Is used by marima, in case there can
be one or more non-identifiable (ar-)parameters when estimating
the lm-object.
}

