% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.forecasting.R
\name{forec.var}
\alias{forec.var}
\title{forec.var}
\usage{
forec.var(marima, nstep = 1, dif.poly = NULL)
}
\arguments{
\item{marima}{marima object (cov.u and ar.estimates and
ma.estimates are used)}

\item{nstep}{length of forecast}

\item{dif.poly}{autoregressive representation of differencing
polynomial as constructed by the function
define.dif(...) when the time series is differenced
(if so) before being analysed by marima.}
}
\value{
pred.var    = variance-covariances for nstep forecasts
(an array with dimension (kvar, kvar, nstep).

rand.shock   = corresponding random shock representation
of the model used.
}
\description{
Function for calculation of variances of nstep forecasts
using a marima type model.
}

